Reorganize chart components and move files into structured folders
