--- docs/prd/15-chart-def-rework/comments.md
@@ -1,5 +1,0 @@
-# Comments on chart def migration
-
-- Needs some work: `src/components/charts/useGetLiveData.tsx`
-- Move axis stuff into folder
-- Move facet stuff into folder
